Reorganize database page
